requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
requests
beautifulsoup4
